backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pypdf==5.1.0
google-generativeai==0.8.4
openai==1.75.0
python-dotenv==1.1.0
pydantic==2.11.3
